Add function prototypes and getChange function
The function prototypes were added for the functions needed to implement
a solution to the minimum change problem. The function prototypes added
were minChange() and printMinChange(). The getChange function uses the
rand() function to return a random value between 1 and 99 that can be
used for the amount of change to be applied to produce the minimum
amount of coins needed to equate to the amount of change.

The getChange function was added to allow for an easy way to produce a
new amount of change for testing the other functions.