Add minChange to find minimum number of coins
The minChange function takes in an int number for the amount of changes
and a pointer called coinArray that holds the amount of each coin needed
for the minimum change. The function changes the values in the coinArray
pointer to hold the coins used for the minimum change.

The minChange function is the main function that allows for it to be
determined how many of each coin needs to be used to use the minimum
number of coins.